edit-one-version

+++ app/Http/Controllers/User/ImageController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\User;
 
-
-use App\Image;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Repositories\Interfaces\ImageRepositoryInterface;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Storage;
 
 class ImageController extends Controller
 {
@@ -21,7 +22,12 @@ class ImageController extends Controller
         $imageId = $request->image;
         $image = $this->imageRepository->findById($imageId);
         $blog = $image->blog;
-        $image->delete();
+        
+        if(File::exists(public_path($image->path))){
+            File::delete(public_path($image->path));
+            $image->delete();       
+        }
+        
         $images = $blog->images;
         return response()->json(['images'=>json_encode($images,true)]);
